Move version to own package and use named arguments
