+++ pkg/version/version.go
@@ -1,4 +1,4 @@
-package pkg
+package version
 
 // VERSION contains version information
 var VERSION = "undefined"
